fixed the many basic broken things about the chat demo.
use a stateful comet actor to hold the user id. the input must be
part of the comet actor so messages can be associated to a user.
blank messages get dropped. all messages are trimmed (spaces would
only matter if they were wrapped in a <pre>, which they aren't).
the whole chat history isn't resent each time, only the new messages.
the template is much cleaner and open to tweaking by web designers.