add /clear command. make session per-tab instead of per-browser.
there's a bug in there someplace, having to do with the data cached
in the javascript side. occasional throws hinting that a doodles is
being destroyed twice. not that motivated right now to hunt it down.
i want to rewrite the javascript delta because it is getting too
convoluted, and maybe that will fix this bug.